The name ShopSimple is an app that makes                                                                                                                       your shopping experience simpler.
The name ShopSimple is a symbol of simplicity as our webpage make things easier for the user

The Html  which is made is a definition of how a simplistic shopping website must be, where it is easy to navigate and has less obstruction such as “ADS”

In the homepage of the website u can see buttons for “About us” and “Contact Us” which make it easy and doesn’t waste ur precious time in understand how to use a website

The Shortcut bar which consists of the icons of “Mobile Phone” and “Electronics” makes it easier for the user to browse around for the phone and its accessories in general.

The content bar below the shortcut bar is a “Quick access panel’ of the based on the Shortcuts in the shortcut bar, so the user can have an overview of the products available

This website will redirect you on a click of a button to flipkart or amazon or any other website, if u want to buy that product.

The Homepage including every other page of the website is only made with “HTML” and “CSS” ; it doesn't consist of any JAVAScript(PHP, node js).